+++ orthanc_tools/dicom_migrator.py
@@ -234,7 +234,7 @@ class DicomMigrator:
             logger.info("To itself")
 
         if self._scheduler:
-            logger.info("Night & Week-end mode Enabled : " + str(self._scheduler._run_only_at_night_and_weekend))
+            logger.info("Scheduler : " + str(self._scheduler))
 
         logger.info("Migrating with {n} threads".format(n=self._worker_threads_count))